LayoutTests/http/tests/websocket/tests/hybi/handshake-fail-by-extensions-header.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../../../../js-test-resources/js-test-pre.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script>
description("Test whether WebSocket handshake fails if the server sends Sec-WebSocket-Extensions header.");

window.jsTestIsAsync = true;
if (window.layoutTestController)
    layoutTestController.overridePreference("WebKitHixie76WebSocketProtocolEnabled", 0);

var url = "ws://localhost:8880/websocket/tests/hybi/handshake-fail-by-extensions-header";
var ws = new WebSocket(url);
var closeEvent;

ws.onopen = function()
{
    testFailed("Connection established.");
    ws.close();
};

ws.onclose = function(event)
{
    closeEvent = event;
    shouldBeFalse("closeEvent.wasClean");
    finishJSTest();
};

</script>
<script src="../../../../js-test-resources/js-test-post.js"></script>
</body>
</html>
